+++ Example/PrimerSDK/CreateClientToken.swift
@@ -1,0 +1,14 @@
+//
+//  CreateClientToken.swift
+//  PrimerSDK_Example
+//
+//  Created by Carl Eriksson on 08/04/2021.
+//  Copyright © 2021 CocoaPods. All rights reserved.
+//
+
+import Foundation
+
+
+struct CreateClientTokenRequest: Codable {
+    let customerId: String
+}
