+++ Example/PrimerSDK/SpinnerViewController.swift
@@ -1,0 +1,25 @@
+//
+//  SpinnerViewController.swift
+//  PrimerSDK_Example
+//
+//  Created by Carl Eriksson on 08/04/2021.
+//  Copyright © 2021 CocoaPods. All rights reserved.
+//
+
+import UIKit
+
+class SpinnerViewController: UIViewController {
+    var spinner = UIActivityIndicatorView(activityIndicatorStyle: .whiteLarge)
+    
+    override func loadView() {
+        view = UIView()
+        view.backgroundColor = UIColor(white: 0, alpha: 0.7)
+        
+        spinner.translatesAutoresizingMaskIntoConstraints = false
+        spinner.startAnimating()
+        view.addSubview(spinner)
+        
+        spinner.centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
+        spinner.centerYAnchor.constraint(equalTo: view.centerYAnchor).isActive = true
+    }
+}
