+++ Example/PrimerSDK/UITableViewCell+Extension.swift
@@ -1,0 +1,35 @@
+//
+//  UITableViewCell+Extension.swift
+//  PrimerSDK_Example
+//
+//  Created by Carl Eriksson on 08/04/2021.
+//  Copyright © 2021 CocoaPods. All rights reserved.
+//
+
+import UIKit
+
+extension UITableViewCell {
+
+    func addIcon(_ icon: UIImage?) {
+        let imageView = UIImageView(image: icon)
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        addSubview(imageView)
+        imageView.centerYAnchor.constraint(equalTo: centerYAnchor).isActive = true
+        imageView.heightAnchor.constraint(equalToConstant: 20).isActive = true
+        imageView.widthAnchor.constraint(equalToConstant: 28).isActive = true
+        imageView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 20).isActive = true
+    }
+
+    func addTitle(_ text: String) {
+        let title = UILabel()
+        title.text = text
+        title.translatesAutoresizingMaskIntoConstraints = false
+        title.adjustsFontSizeToFitWidth = false
+        title.lineBreakMode = .byTruncatingTail
+        addSubview(title)
+        title.centerYAnchor.constraint(equalTo: centerYAnchor).isActive = true
+        title.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 60).isActive = true
+        title.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -50).isActive = true
+    }
+
+}
